+++ src/Utils.php
@@ -60,7 +60,7 @@ class Utils
         if ($maxLen === -1) {
             while (!$stream->eof()) {
                 $buf = $stream->read(1048576);
-                if ($buf === '' || $buf === false) {
+                if ($buf === false) {
                     break;
                 }
                 $buffer .= $buf;
@@ -71,7 +71,7 @@ class Utils
         $len = 0;
         while (!$stream->eof() && $len < $maxLen) {
             $buf = $stream->read($maxLen - $len);
-            if ($buf === '' || $buf === false) {
+            if ($buf === false) {
                 break;
             }
             $buffer .= $buf;