+++ tests/AppendStreamTest.php
@@ -159,4 +159,19 @@ class AppendStreamTest extends \PHPUnit_Framework_TestCase
         $s = new AppendStream();
         $this->assertFalse($s->flush());
     }
+
+    public function testCanDetach()
+    {
+        $s = new AppendStream();
+        $this->assertFalse($s->isDetached());
+        $s->detach();
+        $this->assertTrue($s->isDetached());
+    }
+
+    public function testReturnsEmptyMetadata()
+    {
+        $s = new AppendStream();
+        $this->assertEquals([], $s->getMetadata());
+        $this->assertNull($s->getMetadata('foo'));
+    }
 }